added interface for services resolved conflicts with development
