+++ src/main/java/by/exadel/internship/controller/FormController.java
@@ -11,10 +11,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
+
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
