+++ src/main/java/by/exadel/internship/controller/InternshipController.java
@@ -2,13 +2,14 @@ package by.exadel.internship.controller;
 
 
 import by.exadel.internship.dto.internshipDTO.GuestInternshipDTO;
-import by.exadel.internship.interceptor.LoggerInterceptor;
 import by.exadel.internship.service.InternshipService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
-
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.UUID;
@@ -18,6 +19,7 @@ import java.util.UUID;
 @RequestMapping("/internships")
 @Api(tags = "Endpoints for Internship")
 public class InternshipController {
+
     private final InternshipService internshipService;
 
     @GetMapping
