+++ src/main/java/by/exadel/internship/controller/UserController.java
@@ -1,10 +1,15 @@
 package by.exadel.internship.controller;
+
 import by.exadel.internship.dto.UserDTO;
 import by.exadel.internship.service.UserService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import java.util.List;
 import java.util.UUID;
