+++ src/main/java/by/exadel/internship/service/InternshipService.java
@@ -1,52 +1,13 @@
 package by.exadel.internship.service;
 
-
 import by.exadel.internship.dto.internshipDTO.GuestInternshipDTO;
-import by.exadel.internship.entity.Internship;
-import by.exadel.internship.exception_handing.NotFoundException;
-import by.exadel.internship.mapper.InternshipMapper;
-import by.exadel.internship.repository.InternshipRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.slf4j.MDC;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class InternshipService {
+public interface InternshipService {
 
-    private final InternshipRepository internshipRepository;
-    private final InternshipMapper internShipMapper;
+    GuestInternshipDTO getById(UUID uuid);
+    List<GuestInternshipDTO> getAll();
 
-    public GuestInternshipDTO getById(UUID uuid) {
-
-        MDC.put("className", InternshipService.class.getSimpleName());
-        log.info("Try to get Internships with uuid=" + uuid);
-
-        Internship internship = internshipRepository
-                .findById(uuid)
-                .orElseThrow(() -> new NotFoundException("No such Internship with uuid = " + uuid + " in DB", "uuid.invalid"));
-
-        log.info("Successfully got Internships with uuid=" + uuid);
-
-        return internShipMapper.toGuestInternshipDTO(internship);
-    }
-
-    public List<GuestInternshipDTO> getAll() {
-
-        MDC.put("className", InternshipService.class.getSimpleName());
-        log.info("Try to get all Internships");
-
-        List<Internship> internships = internshipRepository.findAll();
-        List<GuestInternshipDTO> guestInternshipDTOList = internShipMapper.map(internships);
-
-        log.info("Successfully list of Internships");
-
-        return guestInternshipDTOList;
-
-    }
 }
