+++ src/main/java/by/exadel/internship/service/UserService.java
@@ -1,34 +1,13 @@
 package by.exadel.internship.service;
+
 import by.exadel.internship.dto.UserDTO;
-import by.exadel.internship.entity.User;
-import by.exadel.internship.exception.NotFoundException;
-import by.exadel.internship.mapper.UserMapper;
-import by.exadel.internship.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.UUID;
 
-@Service
-@RequiredArgsConstructor
-public class UserService {
+public interface UserService {
 
-    private final UserRepository userRepository;
-    private final UserMapper userMapper;
+    List<UserDTO> getAll();
+    UserDTO getById(UUID uuid);
 
-    public List<UserDTO> getAll() {
-
-        List<User> userList = userRepository.findAll();
-
-        return userMapper.map(userList);
-
-    }
-
-    public UserDTO getById(UUID uuid) {
-
-        User user = userRepository.findById(uuid)
-                .orElseThrow(() -> new NotFoundException("User with id " + uuid + " not found"));
-
-        return userMapper.toUserDTO(user);
-    }
 }
