+++ src/main/java/by/exadel/internship/service/impl/UserServiceImpl.java
@@ -1,0 +1,37 @@
+package by.exadel.internship.service.impl;
+import by.exadel.internship.dto.UserDTO;
+import by.exadel.internship.entity.User;
+import by.exadel.internship.exception.NotFoundException;
+import by.exadel.internship.mapper.UserMapper;
+import by.exadel.internship.repository.UserRepository;
+import by.exadel.internship.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import java.util.List;
+import java.util.UUID;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+    private final UserMapper userMapper;
+
+    @Override
+    public List<UserDTO> getAll() {
+
+        List<User> userList = userRepository.findAll();
+
+        return userMapper.map(userList);
+
+    }
+
+    @Override
+    public UserDTO getById(UUID uuid) {
+
+        User user = userRepository.findById(uuid)
+                .orElseThrow(() -> new NotFoundException("User with id " + uuid + " not found"));
+
+        return userMapper.toUserDTO(user);
+    }
+}
